fix(suite-native): android navigation bar color on switching apps
