+++ packages/styles/src/index.ts
@@ -8,3 +8,4 @@ export * from './mergeStyleObjects';
 export * from './mergeStyles';
 export * from './breakpoints';
 export * from './mediaQueries';
+export * from './utils';
